contributors/gadams/incoming/cssom/cssstyledeclaration-csstext-000.xht

<!DOCTYPE html PUBLIC "-//W3C//DTD//XHTML 1.0 Strict//EN" "http://www.w3.org/TR/xhtml1/DTD/xhtml1-strict.dtd">
<html xmlns="http://www.w3.org/1999/xhtml">
  <head>
    <title>CSSOM Serialization: getting cssText must return the result of serializing the CSS declaration blocks.</title>
    <link rel="author" title="Glenn Adams" href="mailto:glenn@skynav.com"/>
    <link rel="help" href="http://www.w3.org/TR/cssom/#the-cssstyledeclaration-interface"/>
    <meta name="flags" content="dom"/>
    <script src="/resources/testharness.js"/>
    <script src="/resources/testharnessreport.js"/>
  </head>
  <body>
    <div id="log"></div>
    <div id="box"></div>
    <script>
      var style = document.getElementById('box').style;

      test(function(){

        style.left = "10px";
        assert_equals(style.cssText, "left: 10px; ");

      }, 'CSSStyleDeclaration.cssText - delimiters - one declaration.');

      test(function(){

        style.left = "10px";
        style.right = "20px";
        assert_equals(style.cssText, "left: 10px; right: 20px; ");

      }, 'CSSStyleDeclaration.cssText - delimiters - two declarations.');
    </script>
  </body>
</html>
